Rest entities and SocketReactionAddEvent

+++ src/KaiHeiLa.Net.Core/Entities/Channels/IChannel.cs
@@ -13,12 +13,12 @@ public interface IChannel : IULongEntity
     /// <summary>
     ///     创建者 ID
     /// </summary>
-    uint CreateUserId { get; }
+    ulong CreateUserId { get; }
     
     /// <summary>
     ///     Gets a user in this channel.
     /// </summary>
-    /// <param name="id">The snowflake identifier of the user (e.g. <c>168693960628371456</c>).</param>
+    /// <param name="id">The identifier of the user (e.g. <c>168693960628371456</c>).</param>
     /// <param name="mode">The <see cref="CacheMode" /> that determines whether the object should be fetched from cache.</param>
     /// <param name="options">The options to be used when sending the request.</param>
     /// <returns>
